feat: move building of ProjectExport to a separate class
